+++ sdk/schemaregistry/schema-registry-avro/src/models.ts
@@ -4,11 +4,11 @@
 /**
  * A message that contains binary data and a content type.
  */
-export interface MessageWithMetadata {
+export interface MessageContent {
   /**
    * The message's binary data
    */
-  body: Uint8Array;
+  data: Uint8Array;
   /**
    * The message's content type
    */
@@ -23,11 +23,11 @@ export interface MessageAdapter<MessageT> {
   /**
    * defines how to create a message from a payload and a content type
    */
-  produceMessage: (messageWithMetadata: MessageWithMetadata) => MessageT;
+  produceMessage: (messageWithMetadata: MessageContent) => MessageT;
   /**
    * defines how to access the payload and the content type of a message
    */
-  consumeMessage: (message: MessageT) => MessageWithMetadata;
+  consumeMessage: (message: MessageT) => MessageContent;
 }
 
 /**